requirements.txt
django
djangorestframework
django-redis
python-decouple
mysqlclient